+++ src/Ui5/AbstractUi5Report.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace LaravelUi5\Core\Ui5;
+
+use LaravelUi5\Core\Enums\ArtifactType;
+use LaravelUi5\Core\Ui5\Contracts\Ui5ModuleInterface;
+use LaravelUi5\Core\Ui5\Contracts\Ui5ReportInterface;
+
+abstract class AbstractUi5Report implements Ui5ReportInterface
+{
+    public function __construct(protected Ui5ModuleInterface $module)
+    {
+    }
+
+    public function getModule(): Ui5ModuleInterface
+    {
+        return $this->module;
+    }
+
+    public function getType(): ArtifactType
+    {
+        return ArtifactType::Report;
+    }
+}